+++ lib/src/log/LogMessage.dart
@@ -37,7 +37,7 @@ class LogMessage {
 
   Map<String, dynamic> toJson() {
     return <String, dynamic>{
-      'time': time,
+      'time': time.toIso8601String(),
       'source': source,
       'level': level,
       'correlation_id': correlation_id,
@@ -47,7 +47,7 @@ class LogMessage {
   }
 
   void fromJson(Map<String, dynamic> json) {
-    time = json['time'];
+    time = DateTime.parse(json['time']);
     source = json['source'];
     level = json['level'];
     correlation_id = json['correlation_id'];