+++ changelog.md
@@ -5,3 +5,11 @@
 ## 1.0.2
 
 - Fix registerAsType, create and canCreate method in Factory class.
+
+## 1.0.3
+
+- Fix access level to field in CachedLogger class.
+
+## 1.0.4
+
+- Fix json converter in LogMessage class.